Disabled required_approving_review_count in README
Commented out `required_approving_review_count` and added a comment on it. The API is currently in preview for this, and thus breaks setting branch protections.

I wasn't sure if I should remove it entirely or comment it out given the comment about the teams above. If it makes more sense to remove it instead, I'm happy to change it.