feat: Merge arrays of settings by the 'name' attribute
probot-config 1.0.0 will deep-merge the entire settings.yml file. This is
great for settings like 'repository', which do not contain an array.

For arrays, the default behavior is to concatenate the repo's array with
the base array. This is not ideal.

This change will merge any element with a matching name, and will
otherwise concatenate.