fix(direct-collaborators): limited the requested list to direct
since listing all does not align well with the goals of this configuration. `direct` might also be
reasonable option, but this seems like the most likely fit (see
https://developer.github.com/v3/repos/collaborators/#list-collaborators)

BREAKING CHANGE: repositories relying on comparing to the list of all collaborators will no longer
sync the same way, but that use case seems very unlikely. please open an issue to let us know if you
were relying on this behavior

fixes #155